Fix linters and remove getmeals.js file
